+++ app/Http/Controllers/OrderQuotationController.php
@@ -222,8 +222,10 @@ class OrderQuotationController extends Controller
                 $option = ApprovalOption::findByType($user->company_id, ApprovalType::ApprovalQuotation->value);
 
                 $approvalOrder = ApprovalOrder::where('order_id', $order->id)->where('approval_type', $option->type)->first();
-                ApprovalOrderProcess::where('approval_order_id', $approvalOrder->id)->delete();
-                $approvalOrder->delete();
+                if ($approvalOrder) {
+                    ApprovalOrderProcess::where('approval_order_id', $approvalOrder->id)->delete();
+                    $approvalOrder->delete();
+                }
 
                 $approvalOrder = ApprovalOrder::create([
                     'order_id' => $order->id,
