+++ app/Http/Controllers/ProviderQuotationController.php
@@ -183,8 +183,10 @@ class ProviderQuotationController extends Controller
             $approvers = $option->approver;
 
             $approvalOrder = ApprovalOrder::where('order_id', $order->id)->where('approval_type', $option->type)->first();
-            ApprovalOrderProcess::where('approval_order_id', $approvalOrder->id)->delete();
-            $approvalOrder->delete();
+            if ($approvalOrder) {
+                ApprovalOrderProcess::where('approval_order_id', $approvalOrder->id)->delete();
+                $approvalOrder->delete();
+            }
 
             $approvalOrder = ApprovalOrder::create([
                 'order_id' => $order->id,
